+++ src/main/java/blusunrize/immersiveengineering/common/items/BuzzsawItem.java
@@ -32,6 +32,8 @@ import blusunrize.immersiveengineering.common.util.ItemNBTHelper;
 import blusunrize.immersiveengineering.common.util.Utils;
 import blusunrize.immersiveengineering.common.util.fluids.IEItemFluidHandler;
 import blusunrize.immersiveengineering.common.util.inventory.IEItemStackHandler;
+import com.google.common.collect.ImmutableMultimap;
+import com.google.common.collect.ImmutableMultimap.Builder;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multimap;
 import net.minecraft.block.Block;
@@ -356,20 +358,20 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 	@Override
 	public Multimap<Attribute, AttributeModifier> getAttributeModifiers(EquipmentSlotType slot, ItemStack stack)
 	{
-		Multimap<Attribute, AttributeModifier> multimap = super.getAttributeModifiers(slot, stack);
+		Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
 		if(slot==EquipmentSlotType.MAINHAND)
 		{
 			ItemStack sawblade = getSawblade(stack);
 			if(!sawblade.isEmpty()&&canBuzzsawBeUsed(stack, null))
 			{
-				multimap.put(Attributes.ATTACK_DAMAGE,
+				builder.put(Attributes.ATTACK_DAMAGE,
 						new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Tool modifier",
 								((SawbladeItem)sawblade.getItem()).getSawbladeDamage(), Operation.ADDITION));
-				multimap.put(Attributes.ATTACK_SPEED,
+				builder.put(Attributes.ATTACK_SPEED,
 						new AttributeModifier(ATTACK_SPEED_MODIFIER, "Tool modifier", -2.5D, Operation.ADDITION));
 			}
 		}
-		return multimap;
+		return builder.build();
 	}
 
 	@Override