+++ changelog.md
@@ -1,6 +1,7 @@
 ##### Version 0.16-109
 - Fixed recipes not loading on dedicated servers (BluSunrize)
 - Fixed fluid tank backgrounds being wrongly offset in JEI (BluSunrize)
+- Fixed revolver causing a crash with its attribute map (BluSunrize)
 - Also includes the changes from 0.14-107
 
 ##### Version 0.15-108